Remove --babelrc, cause Babel to use non-archetype config
This removes `--babelrc` and will require boilerplates and other
projects using this archetype to add their own `.babelrc` if they
want non-default settings.

This PR does not remove the actual `.babelrc` file in this repo – it's
not hurting anything, can be used as a reference, and will most likely
be used again later when custom babelrc is more widely supported.

Reference: #4 (not yet fixed)

/cc @boygirl @ryan-roemer